requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blis==1.3.3
bs4==0.0.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
courlan==1.3.2
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.13
dateparser==1.2.2
dirtyjson==1.0.8
distro==1.9.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
exceptiongroup==1.3.1
faiss-cpu==1.13.1
fastapi==0.123.7
filelock==3.20.0
fonttools==4.61.1
fqdn==1.5.1
fsspec==2025.12.0
graphql-core==3.2.7
h11==0.16.0
harfile==0.4.0
hf-xet==1.2.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hypothesis==6.148.6
hypothesis-graphql==0.11.1
hypothesis-jsonschema==0.23.1
idna==3.11
iniconfig==2.3.0
isoduration==20.11.0
jaxtyping==0.3.4
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
junit-xml==1.9
jusText==3.0.2
kaggle==1.7.4.5
kiwisolver==1.4.9
langdetect==1.0.9
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.15
narwhals==2.13.0
networkx==3.4.2
numpy==2.2.6
ollama==0.6.1
openai==2.8.1
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
plotly==6.5.0
pluggy==1.6.0
preshed==3.0.12
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pypdfium2==5.1.0
pyrate-limiter==3.9.0
pysbd==0.3.4
pytest==9.0.1
pytest-subtests==0.14.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==14.2.0
rpds-py==0.30.0
safetensors==0.7.0
schemathesis==4.6.6
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.2
starlette==0.50.0
starlette-testclient==0.4.1
sympy==1.14.0
tenacity==9.1.2
termplotlib==0.3.9
text-unidecode==1.3
thinc==8.3.10
threadpoolctl==3.6.0
tld==0.13.1
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.57.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
wadler_lindig==0.1.7
wasabi==1.1.3
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
Werkzeug==3.1.4
wrapt==2.0.1
